refactor(ui): promote TintedIcon to a shared primitive
Move TintedIcon out of toasts into ui/components with its own tint styling and an optional tint override. Toast consumes it from the barrel; drop the dead iconTint style.
