+++ packages/client/src/ui/components/index.ts
@@ -1,2 +1,3 @@
+export { default as TintedIcon } from './tinted-icon';
 export { default as Empty } from './empty-state';
 export { default as Page } from './page';
