+++ packages/client/src/ui/toasts/index.ts
@@ -3,6 +3,5 @@ export { default as useToastState } from './use-toast-state';
 export { default as ToastProgress } from './toast-progress';
 export { default as ToastButtons } from './toast-buttons';
 export { default as ToastContent } from './toast-content';
-export { default as TintedIcon } from './tinted-icon';
 export { default as ToastIcon } from './toast-icon';
 export { default as Toast } from './toast';
